+++ src/main/java/blackjack/domain/Result.java
@@ -1,5 +1,6 @@
 package blackjack.domain;
 
+import blackjack.domain.player.Gamer;
 import java.util.Arrays;
 import java.util.function.BiPredicate;
