+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -1,5 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
 import java.util.ArrayList;
 import java.util.List;
