+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
+import blackjack.domain.card.Card;
 import java.util.List;
 
 public interface Player {
