+++ src/main/java/blackjack/view/OutputView.java
@@ -2,9 +2,9 @@ package blackjack.view;
 
 import static java.util.stream.Collectors.joining;
 
-import blackjack.domain.Card;
-import blackjack.domain.Dealer;
-import blackjack.domain.Player;
+import blackjack.domain.card.Card;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Player;
 import blackjack.domain.Result;
 import java.util.EnumMap;
 import java.util.List;
